target/rivanna/requirements.txt
#tensorflow==2.9.0
# protobuf==3.20.1
grpcio
numpy==1.23.5
requests==2.27.1
smartredis==0.3.1
# tensorflow==2.11.0
tensorflow==2.13.0
tensorflow-serving-api
tqdm==4.64.0
cloudmesh-ee
cloudmesh-common
cloudmesh-gpu
cloudmesh-rivanna
pyyaml
matplotlib
docopt
# git+https://github.com/mlperf/logging.git@2.1.0
# git+https://github.com/mlperf/logging.git@1.0.0
